third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-integer-expected.txt

This is a testharness.js-based test.
PASS Animating a custom property of type <integer>
PASS Animating a custom property of type <integer> with a single keyframe
PASS Animating a custom property of type <integer> with additivity
PASS Animating a custom property of type <integer> with a single keyframe and additivity
FAIL Animating a custom property of type <integer> with iterationComposite assert_equals: expected "250" but got "50"
Harness: the test ran to completion.
